requirements.txt
asgiref==3.5.2
certifi==2022.6.15
decorator==4.4.2
distlib==0.3.5
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
django-model-utils==4.0.0
django-split-settings==1.1.0
djangorestframework==3.13.1
filelock==3.8.0
gunicorn==20.0.4
Markdown==3.4.1
pipenv==2022.8.19
platformdirs==2.5.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
validators==0.18.1
virtualenv==20.16.3
virtualenv-clone==0.5.7
